refactor(pdata): share batch sizing policy
Move the common batch sizing enum and payload measurement logic into pdata so receivers and processors use one definition.
